% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dd_db_connect}
\alias{dd_db_connect}
\title{Connect to the dundee SQLite store.}
\usage{
dd_db_connect(cfg, create = TRUE)
}
\arguments{
\item{cfg}{A config list (see \code{\link[=dd_config]{dd_config()}}).}

\item{create}{Logical; create parent directory if missing.}
}
\value{
A DBIConnection. Caller is responsible for \code{DBI::dbDisconnect()}.
}
\description{
Connect to the dundee SQLite store.
}
