% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dd_db_init}
\alias{dd_db_init}
\title{Initialise (or migrate to) the dundee schema. Idempotent.}
\usage{
dd_db_init(con)
}
\arguments{
\item{con}{A DBIConnection from \code{\link[=dd_db_connect]{dd_db_connect()}}.}
}
\value{
\code{con}, invisibly.
}
\description{
Initialise (or migrate to) the dundee schema. Idempotent.
}
