% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dd_config_example}
\alias{dd_config_example}
\title{Write an example configuration file.}
\usage{
dd_config_example(path = "config.example.yml")
}
\arguments{
\item{path}{Destination path for the example YAML.}
}
\value{
\code{path}, invisibly.
}
\description{
Write an example configuration file.
}
